+++ lib/lightning_web/live/workflow_live/index.ex
@@ -5,10 +5,10 @@ defmodule LightningWeb.WorkflowLive.Index do
   import LightningWeb.WorkflowLive.Components
 
   alias Lightning.DashboardStats
-  alias Lightning.Extensions.RuntimeLimiter
-  alias Lightning.Extensions.RuntimeLimiting.Context
+  alias Lightning.Extensions.UsageLimiting.Context
   alias Lightning.Policies.Permissions
   alias Lightning.Policies.ProjectUsers
+  alias Lightning.Services.UsageLimiter
   alias Lightning.Workflows
   alias LightningWeb.WorkflowLive.DashboardComponents
   alias LightningWeb.WorkflowLive.NewWorkflowForm
@@ -88,7 +88,7 @@ defmodule LightningWeb.WorkflowLive.Index do
       NewWorkflowForm.validate(%{}, socket.assigns.project.id)
     )
     |> then(fn socket ->
-      case RuntimeLimiter.check_limits(%Context{
+      case UsageLimiter.check_limits(%Context{
              project_id: socket.assigns.project.id
            }) do
         :ok ->