+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to
 
 ### Added
 
+- Add RunQueue extension to allow claim customization.
+  [#1715](https://github.com/OpenFn/Lightning/issues/1715)
+
 ### Changed
 
 - Use `PAYLOAD_SIZE_KB` in k6 load testing script, set thresholds on wait time,